// libs/components/product/Review.tsx
import React from 'react';
import { Stack, Typography } from '@mui/material';
import { Comment } from '../../types/comment/comment';
import { API_URL } from '../../config';
import { format } from 'date-fns';
import { useRouter } from 'next/router';
import { useReactiveVar } from '@apollo/client';
import { userVar } from '../../../apollo/store';

interface ReviewProps {
	comment: Comment;
}

const Review = (props: ReviewProps) => {
	const { comment } = props;
	const router = useRouter();
	const user = useReactiveVar(userVar);
	const imagePath: string = comment?.memberData?.memberImage
		? `${API_URL}/${comment?.memberData?.memberImage}`
		: '/img/profile/defaultUser.svg';

	/** HANDLERS **/
	const goMemberPage = (id: string) => {
		if (id === user?._id) router.push('/mypage');
		else router.push(`/member?memberId=${id}`);
	};

	return (
		<Stack className={'review-config'}>
			<Stack className={'review-mb-info'}>
				<Stack className={'img-name-box'}>
					<img src={imagePath} alt="" className={'img-box'} />
					<Stack>
						<Typography
							className={'name'}
							onClick={() => goMemberPage(comment?.memberData?._id as string)}
						>
							{comment.memberData?.memberNick}
						</Typography>
						<Typography className={'date'}>
							{format(new Date(comment.createdAt), 'dd MMMM, yyyy')}
						</Typography>
					</Stack>
				</Stack>
			</Stack>
			<Stack className={'desc-box'}>
				<Typography className={'description'}>{comment.commentContent}</Typography>
			</Stack>
		</Stack>
	);
};

export default Review;
